Consolidate tooltip components to fix duplicate identifier errors
- Replaced all core/ui/Tooltip imports with base-v2/ui/tooltip
- Added TooltipProvider wrappers to all tooltip usages
- Removed old core/ui/Tooltip.tsx file to eliminate conflicts
- Fixes TypeScript compilation errors caused by duplicate Tooltip exports

This resolves the TypeScript build errors by ensuring only one tooltip implementation is used throughout the application.
